separate plugin and non-plugin tests
